edit type error add folder error model
