+++ src/components/UserProvider/userProvider.type.ts
@@ -1,3 +1,4 @@
+import { Error } from 'models/error';
 import { User } from 'models/user';
 
 export interface BaseResponse<T> {
@@ -8,18 +9,6 @@ export interface BaseResponse<T> {
 export interface LoginRequestPayload {
   username: string;
   password: string;
-}
-export interface ErrorBase<T> {
-  statusCode: number;
-  name: string;
-  message: T;
-  timestamp: Date;
-  path: string;
-}
-export interface Error<T> {
-  response: {
-    data: ErrorBase<T>;
-  };
 }
 
 export type LoginSuccessPayload = BaseResponse<string>;
